cfgVehicleNew.unlockConditionCount int -> bigint

+++ logic-server/src/main/java/logic/server/dto/CfgVehicleNewDTO.java
@@ -14,7 +14,7 @@ public class CfgVehicleNewDTO implements DTO{
     private int preConditionChallengeType;
     private int preConditionChallengeId;
     private int unlockConditionType;
-    private int unlockConditionCount;
+    private long unlockConditionCount;
     private int showIndex;
     private int levelMax;
     private String levelUpFormula;
